% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_naive.R
\name{grad01_naive_h}
\alias{grad01_naive_h}
\title{Calculate Gradient (Naive Method, H2/S2 Parameterization, Conditional Model)}
\usage{
grad01_naive_h(par, rho, y, K1_list, K2_list, i1, i0)
}
\arguments{
\item{par}{A numeric vector of parameters (h21, h22, ..., h2M) for the
parameters being optimized under the null hypothesis. The preceding elements are
proportions of variance (\code{h2_i}) for \code{K2_list}. The the total variance \code{s2} has closed form estimation.}

\item{rho}{A numeric vector of fixed \code{h2} values (h2_1, h2_2, ...) under the null
hypothesis. These correspond to \code{K1_list}.}

\item{y}{Numeric vector of observed data for the whole subset.}

\item{K1_list}{A list of full covariance matrices corresponding to the first set of \code{h2_i} parameters.}

\item{K2_list}{A list of full covariance matrices corresponding to the second set of \code{h2_i} parameters.}

\item{i1}{Indices used to subset the full covariance matrices for Y1.}

\item{i0}{Indices used to subset the full covariance matrices for Y0.}
}
\value{
A numeric vector of gradients corresponding to each parameter in \code{par}.
}
\description{
This function computes the gradient of the log-likelihood for the conditional Model (Y0|Y1)
using a "naive" approach with full covariance matrices and an h2/s2 parameterization.
\code{par} is ordered as (h21, h22, ..., h2M).
}
