% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ortho.R
\name{grad0_ortho_h}
\alias{grad0_ortho_h}
\title{Calculate Gradient (Ortho Method, H2/S2 Parameterization, Conditional Model)}
\usage{
grad0_ortho_h(par, rho, y, K_list, TOL = 1e-08)
}
\arguments{
\item{par}{A numeric vector of parameters (h21, h22, ..., h2M) for the
parameters being optimized under the null hypothesis. The preceding elements are
proportions of variance (\code{h2_i}) for \code{K2_list}. The the total variance \code{s2} has closed form estimation.}

\item{rho}{A numeric vector of fixed \code{h2} values (h2_1, h2_2, ...) under the null
hypothesis. These correspond to \code{K1_list}.}

\item{y}{Numeric vector of observed data for the Y0 subset.}

\item{K_list}{A list of numeric vectors, where each \code{K_list[[i]]} represents the
diagonal values of a transformed covariance matrix.}

\item{TOL}{Numerical tolerance to be considered as zero (default: 1e-8).}
}
\value{
A numeric vector of gradients corresponding to each parameter in \code{par}.
}
\description{
This function computes the gradient of the log-likelihood for the conditional Model (Y0|Y1)
using a "naive" approach with full covariance matrices and an h2/s2 parameterization.
\code{par} is ordered as (h21, h22, ..., h2M).
}
