% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_naive.R
\name{grad1_naive_h}
\alias{grad1_naive_h}
\title{Calculate Gradient (Naive Method, H2/S2 Parameterization, Full Model)}
\usage{
grad1_naive_h(par, y1, K1_list, K2_list, i1)
}
\arguments{
\item{par}{A numeric vector of parameters (h21, h22, ..., h2M). The preceding elements are
proportions of variance (\code{h2_i}). The total variance \code{s2} has closed form estimation.}

\item{y1}{Numeric vector of observed data for the Y1 subset.}

\item{K1_list}{A list of full covariance matrices corresponding to the first set of \code{h2_i} parameters.}

\item{K2_list}{A list of full covariance matrices corresponding to the second set of \code{h2_i} parameters.}

\item{i1}{Indices used to subset the full covariance matrices for Y1.}
}
\value{
A numeric vector of gradients corresponding to each parameter in \code{par}.
}
\description{
This function computes the gradient of the log-likelihood for the full model (Y1)
using a "naive" approach with full covariance matrices and an h2/s2 parameterization.
\code{par} is ordered as (h21, h22, ..., h2M).
}
