% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ortho.R
\name{grad1_ortho_h}
\alias{grad1_ortho_h}
\title{Calculate Gradient (ortho Method, H2/S2 Parameterization, Full Model)}
\usage{
grad1_ortho_h(par, y, K_list, TOL = 1e-08)
}
\arguments{
\item{par}{A numeric vector of parameters (h21, h22, ..., h2M). The preceding elements are
proportions of variance (\code{h2_i}). The total variance \code{s2} has closed form estimation.}

\item{y}{Numeric vector of observed data for the Y1 subset.}

\item{K_list}{A list of numeric vectors, where each \code{K_list[[i]]} represents the
diagonal values of a transformed covariance matrix.}

\item{TOL}{Numerical tolerance to be considered as zero (default: 1e-8).}
}
\value{
A numeric vector of gradients corresponding to each parameter in \code{par}.
}
\description{
This function computes the gradient of the log-likelihood for the full model (Y1)
using a "naive" approach with full covariance matrices and an h2/s2 parameterization.
\code{par} is ordered as (h21, h22, ..., h2M).
}
