% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_test0.R
\name{grad1_test0_h}
\alias{grad1_test0_h}
\title{Calculate Gradient (Test for Zero Method, H2/S2 Parameterization, Full Model)}
\usage{
grad1_test0_h(par, y1, K1_list, K2, i1)
}
\arguments{
\item{par}{A numeric vector of parameters (h12, h22, ..., h2M, s2).
\code{par[1:n_K1]} corresponds to \code{K1_list}, \code{par[length(par)-1]} corresponds to \code{K2},
and \code{par[length(par)]} is the total variance \code{s2}.}

\item{y1}{Numeric vector of observed data for the Y1 subset.}

\item{K1_list}{A list of full covariance matrices (pre-transformed by V2) corresponding to the first set of \code{h2_i} parameters.}

\item{K2}{A numeric vector representing the diagonal values of the transformed K2 matrix.}

\item{i1}{Indices used to subset the data and K matrices for Y1.}
}
\value{
A numeric vector of gradients corresponding to each parameter in \code{par}.
}
\description{
This function computes the gradient of the log-likelihood for the full model (Y1)
in the "test for zero" scenario, using h2/s2 parameterization.
}
