% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ortho.R
\name{grad_ortho_s2e}
\alias{grad_ortho_s2e}
\title{Calculate Gradient (Orthogonal K, Sigma2 Parameterization, for s2e)}
\usage{
grad_ortho_s2e(par, y, K_list)
}
\arguments{
\item{par}{The full parameter vector (s21, s22, ..., s2e).}

\item{y}{Numeric vector of observed data.}

\item{K_list}{A list of numeric vectors, where each \code{K_list[[i]]} represents the
diagonal values of a transformed covariance matrix.}
}
\value{
The gradient value for \code{s2e}.
}
\description{
Computes the gradient of the log-likelihood with respect to \code{s2e} (residual error variance)
in the orthogonal K matrices case.
}
