% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ortho.R
\name{grad_ortho_s2i}
\alias{grad_ortho_s2i}
\title{Calculate Gradient (Orthogonal K, Sigma2 Parameterization, for s2_i)}
\usage{
grad_ortho_s2i(s2i_param_index, par, y, K_list, TOL = 1e-08)
}
\arguments{
\item{s2i_param_index}{The index of the specific \code{s2_i} in the \code{par} vector.}

\item{par}{The full parameter vector (s21, s22, ..., s2e).}

\item{y}{Numeric vector of observed data.}

\item{K_list}{A list of numeric vectors, where each \code{K_list[[i]]} represents the
diagonal values of a transformed covariance matrix.}

\item{TOL}{Numerical tolerance to be considered as zero (default: 1e-8).}
}
\value{
The gradient value for \code{s2_i}.
}
\description{
Computes the gradient of the log-likelihood with respect to a specific \code{s2_i}
parameter in the orthogonal K matrices case.
}
