% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_naive.R
\name{loglik01_naive_h}
\alias{loglik01_naive_h}
\title{Calculate Log-Likelihood (Naive Method, H2/S2 Parameterization, Conditional Model)}
\usage{
loglik01_naive_h(par, rho, y0, y1, K1_list, K2_list, i1, i0)
}
\arguments{
\item{par}{A numeric vector of parameters (h2_k, ..., h2_total, s2) for the
parameters being optimized under the null hypothesis. The last element
\code{par[length(par)]} is the total variance \code{s2}. The preceding elements are
proportions of variance (\code{h2_i}) for \code{K2_list}.}

\item{rho}{A numeric vector of fixed \code{h2} values (h2_1, h2_2, ...) under the null
hypothesis. These correspond to \code{K1_list}.}

\item{y0}{Numeric vector of observed data for the Y0 subset.}

\item{y1}{Numeric vector of observed data for the Y1 subset.}

\item{K1_list}{A list of full covariance matrices corresponding to the \code{rho} parameters.}

\item{K2_list}{A list of full covariance matrices corresponding to the \code{par} parameters (excluding s2).}

\item{i1}{Indices used to subset the full covariance matrices for Y1.}

\item{i0}{Indices used to subset the full covariance matrices for Y0.}
}
\value{
The log-likelihood value.
}
\description{
This function computes the log-likelihood for the conditional model (Y0|Y1) using a
"naive" approach with full covariance matrices and an h2/s2 parameterization.
}
