% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_naive.R
\name{loglik01_naive_s}
\alias{loglik01_naive_s}
\title{Calculate Log-Likelihood (Naive Method, Sigma2 Parameterization, Conditional Model)}
\usage{
loglik01_naive_s(par, rho, y0, y1, K1_list, K2_list, i1, i0)
}
\arguments{
\item{par}{A numeric vector of variance component parameters (s21, s22, ..., s2e)
for the parameters being optimized under the null hypothesis (excluding rho).
The last element is assumed to be the residual error variance (s2e).}

\item{rho}{A numeric vector of fixed variance component parameters (s21, s22, ...)
under the null hypothesis. These correspond to \code{K1_list}.}

\item{y0}{Numeric vector of observed data for the Y0 subset.}

\item{y1}{Numeric vector of observed data for the Y1 subset.}

\item{K1_list}{A list of full covariance matrices corresponding to the \code{rho} parameters.}

\item{K2_list}{A list of full covariance matrices corresponding to the \code{par} parameters (excluding s2e).}

\item{i1}{Indices used to subset the full covariance matrices for Y1.}

\item{i0}{Indices used to subset the full covariance matrices for Y0.}
}
\value{
The log-likelihood value.
}
\description{
This function computes the log-likelihood for the conditional model (Y0|Y1) using a
"naive" approach with full covariance matrices and a direct sigma2 parameterization.
}
