% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_test0.R
\name{loglik01_test0_h}
\alias{loglik01_test0_h}
\title{Calculate Log-Likelihood (Test for Zero Method, H2/S2 Parameterization, Conditional Model)}
\usage{
loglik01_test0_h(par, y0, y1, K1_list, K2, i1, i0)
}
\arguments{
\item{par}{A numeric vector of parameters for (h21, h22, ..., h2M, s2). \code{par[1,...,M]} is
the \code{h2} for \code{K2}, and \code{par[length(par)]} is the total variance \code{s2}.}

\item{y0}{Numeric vector of observed data for the Y0 subset.}

\item{y1}{Numeric vector of observed data for the Y1 subset.}

\item{K1_list}{A list of full covariance matrices (pre-transformed by V2) corresponding to \code{rho}.}

\item{K2}{A numeric vector representing the diagonal values of the transformed K2 matrix.}

\item{i1}{Indices used to subset the data and K matrices for Y1.}

\item{i0}{Indices used to subset the data and K matrices for Y0.}
}
\value{
The log-likelihood value.
}
\description{
This function computes the log-likelihood for the conditional model (Y0|Y1) in the
"test for zero" scenario, using h2/s2 parameterization.
}
