% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_test0.R
\name{loglik01_test0_s}
\alias{loglik01_test0_s}
\title{Calculate Log-Likelihood (Test for Zero Method, Sigma2 Parameterization, Conditional Model)}
\usage{
loglik01_test0_s(par, rho, y0, y1, K1_list, K2, i1, i0)
}
\arguments{
\item{par}{A numeric vector of variance component parameters for the \code{K2} matrix
and the residual error (\code{s2e}). \code{par[1]} is the variance component for \code{K2},
and \code{par[length(par)]} is \code{s2e}.}

\item{rho}{A numeric vector of fixed variance component parameters (s21, s22, ...)
under the null hypothesis. These correspond to \code{K1_list}.}

\item{y0}{Numeric vector of observed data for the Y0 subset.}

\item{y1}{Numeric vector of observed data for the Y1 subset.}

\item{K1_list}{A list of full covariance matrices (pre-transformed by V2) corresponding to \code{rho}.}

\item{K2}{A numeric vector representing the diagonal values of the transformed K2 matrix.}

\item{i1}{Indices used to subset the data and K matrices for Y1.}

\item{i0}{Indices used to subset the data and K matrices for Y0.}
}
\value{
The log-likelihood value.
}
\description{
This function computes the log-likelihood for the conditional model (Y0|Y1) in the
"test for zero" scenario, where K2 is assumed to be a single matrix and
its diagonal values are used.
}
