% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ortho.R
\name{loglik0_ortho_h}
\alias{loglik0_ortho_h}
\title{Calculate Log-Likelihood (Orthogonal K, H2/S2 Parameterization)}
\usage{
loglik0_ortho_h(par, rho, y, K_list, s2hat = NULL)
}
\arguments{
\item{par}{A numeric vector of parameters (h21, h22, ..., h2M). The preceding elements are
proportions of variance (\code{h2_i}). The total variance \code{s2} has closed form estimation.}

\item{rho}{A numeric vector of fixed \code{h2} values (h2_1, h2_2, ...) under the null
hypothesis. These correspond to \code{K1_list}.}

\item{y}{Numeric vector of observed data Y0.}

\item{K_list}{A list of numeric vectors, where each \code{K_list[[i]]} represents the
diagonal values of a transformed covariance matrix.}

\item{s2hat}{A numeric value of given estimation of s2. Default as NULL.}
}
\value{
The log-likelihood value.
}
\description{
Computes the log-likelihood for the orthogonal K matrices case with h2/s2 parameterization.
\code{par} is ordered as (h21, h22, ..., h2M).
}
