% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ortho.R
\name{loglik1_ortho_h}
\alias{loglik1_ortho_h}
\title{Calculate Log-Likelihood (Orthogonal K, H2/S2 Parameterization)}
\usage{
loglik1_ortho_h(par, y, K_list, return.s2phat = FALSE)
}
\arguments{
\item{par}{A numeric vector of parameters. \code{par[1]} to \code{par[length(par)-1]}
are \code{h2_i} (proportions of variance for \code{K_list[[i]]}). The last parameter,
\code{par[length(par)]}, is \code{s2} (total variance).}

\item{y}{Numeric vector of observed data Y1.}

\item{K_list}{A list of numeric vectors, where each \code{K_list[[i]]} represents the
diagonal values of a transformed covariance matrix.}

\item{return.s2phat}{Boolean. Determine if estimation of s2 is returned, default at FALSE}
}
\value{
The log-likelihood value.
}
\description{
Computes the log-likelihood for the orthogonal K matrices case with h2/s2 parameterization.
\code{par} is ordered as (h12, h22, ...). s2 has closed form estimation.
}
