% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_test0.R
\name{loglik1_test0_h}
\alias{loglik1_test0_h}
\title{Calculate Log-Likelihood (Test for Zero Method, H2/S2 Parameterization, Full Model)}
\usage{
loglik1_test0_h(par, y1, K1_list, K2, i1, return.s2phat = FALSE)
}
\arguments{
\item{par}{A numeric vector of parameters (h21, h22, ..., h2M). The elements are
proportions of variance (\code{h2_i}). The total variance \code{s2} has closed form estimation.}

\item{y1}{Numeric vector of observed data for the Y1 subset.}

\item{K1_list}{A list of full covariance matrices (pre-transformed by V2) corresponding to the first set of \code{h2_i} parameters.}

\item{K2}{A numeric vector representing the diagonal values of the transformed K2 matrix.}

\item{i1}{Indices used to subset the data and K matrices for Y1.}

\item{return.s2phat}{Boolean. Determine if estimation of s2 is returned, default at FALSE}
}
\value{
The log-likelihood value.
}
\description{
This function computes the log-likelihood for the full model (Y1) in the
"test for zero" scenario, using h2/s2 parameterization. K2 is assumed to be a
single matrix whose diagonal values are used.
}
