% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ortho.R
\name{loglik_ortho_h}
\alias{loglik_ortho_h}
\title{Calculate Log-Likelihood (Orthogonal K, H2/S2 Parameterization)}
\usage{
loglik_ortho_h(par, y, K_list)
}
\arguments{
\item{par}{A numeric vector of parameters. \code{par[1]} to \code{par[length(par)-1]}
are \code{h2_i} (proportions of variance for \code{K_list[[i]]}). The last parameter,
\code{par[length(par)]}, is \code{s2} (total variance).}

\item{y}{Numeric vector of observed data.}

\item{K_list}{A list of numeric vectors, where each \code{K_list[[i]]} represents the
diagonal values of a transformed covariance matrix.}
}
\value{
The log-likelihood value.
}
\description{
Computes the log-likelihood for the orthogonal K matrices case with h2/s2 parameterization.
\code{par} is ordered as (h12, h22, ..., s2).
}
