\name{print.varComp.test}
\alias{print.varComp.test}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Print variance component test results
}
\description{
Currently only the p-values are printed.
}
\usage{
\method{print}{varComp.test}(x, ...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
	An object of class \code{varComp.test}
}
  \item{\dots}{
	Not used.
}
}
\value{
Invisible \code{p.value(x)}
}
\author{
Long Qu
}
\keyword{ print }
