% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_naive.R
\name{ratio_naive_h}
\alias{ratio_naive_h}
\title{Calculate Likelihood Ratio (Naive Method, H2/S2 Parameterization)}
\usage{
ratio_naive_h(y0, y1, K1_list, K2_list, i1, i0, opt1, opt01)
}
\arguments{
\item{y0}{Numeric vector of observed data for the Y0 subset.}

\item{y1}{Numeric vector of observed data for the Y1 subset.}

\item{K1_list}{A list of full covariance matrices corresponding to the parameters that are fixed under the null.}

\item{K2_list}{A list of full covariance matrices corresponding to the parameters that are optimized.}

\item{i1}{Indices used to subset the full covariance matrices for Y1.}

\item{i0}{Indices used to subset the full covariance matrices for Y0.}

\item{opt1}{Optimization result object for the alternative hypothesis (full model).
Must contain \verb{$par} for estimated parameters.}

\item{opt01}{Optimization result object for the null hypothesis. Must contain \verb{$value} for the log-likelihood.}
}
\value{
The likelihood ratio (exp(L_theta1 - L_rho)).
}
\description{
Computes the likelihood ratio for the naive method with h2/s2 parameterization.
}
