% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_ortho.R
\name{ratio_ortho_s}
\alias{ratio_ortho_s}
\title{Calculate Likelihood Ratio (Orthogonal K, Sigma2 Parameterization)}
\usage{
ratio_ortho_s(y, K_list, opt1, opt01)
}
\arguments{
\item{y}{Numeric vector of observed data of Y0.}

\item{K_list}{A list of numeric vectors, where each \code{K_list[[i]]} represents the
diagonal values of a transformed covariance matrix, corresponding to \code{par[i]}.}

\item{opt1}{Optimization result object for the alternative hypothesis (full model).
Must contain \verb{$par} for estimated parameters.}

\item{opt01}{Optimization result object for the null hypothesis. Must contain \verb{$value} for the log-likelihood.}
}
\value{
The likelihood ratio (exp(L_theta1 - L_rho)).
}
\description{
Computes the likelihood ratio for the orthogonal method with sigma2 parameterization.
}
