% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_test0.R
\name{ratio_test0_h}
\alias{ratio_test0_h}
\title{Calculate Likelihood Ratio (Test for Zero Method, H2/S2 Parameterization)}
\usage{
ratio_test0_h(y0, y1, K1_list, K2, i1, i0, opt1, s2hat1, opt01)
}
\arguments{
\item{y0}{Numeric vector of observed data for the Y0 subset.}

\item{y1}{Numeric vector of observed data for the Y1 subset.}

\item{K1_list}{A list of full covariance matrices (pre-transformed by V2) corresponding to parameters fixed under the null.}

\item{K2}{A numeric vector representing the diagonal values of the transformed K2 matrix.}

\item{i1}{Indices used to subset the data and K matrices for Y1.}

\item{i0}{Indices used to subset the data and K matrices for Y0.}

\item{opt1}{Optimization result object for the alternative hypothesis (full model).
Must contain \verb{$par} for estimated parameters.}

\item{s2hat1}{Numeric scaler as estimation of s2 based on Y1.}

\item{opt01}{Optimization result object for the null hypothesis. Must contain \verb{$value} for the log-likelihood.}
}
\value{
The likelihood ratio (exp(L_theta1 - L_rho)).
}
\description{
Computes the likelihood ratio for the "test for zero" method with h2/s2 parameterization.
}
