% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_test0.R
\name{slrt_test0}
\alias{slrt_test0}
\title{Split Likelihood Ratio Test for multiple variance components (Test for Zero Method)}
\usage{
slrt_test0(y, K1_list, K2_list, i1, i0, parameter.set = "h2", TOL = 1e-08)
}
\arguments{
\item{y}{Numeric vector of observed data.}

\item{K1_list}{A list of full covariance matrices corresponding to parameters
to be tested (fixed at 0 under the null). These will be transformed by V2.}

\item{K2_list}{A list containing a single full covariance matrix (K2) which is the nuisance parameter.
This matrix will be eigen-decomposed.}

\item{i1}{Indices for the Y1 subset of data (used in the alternative hypothesis).}

\item{i0}{Indices for the Y0 subset of data (used in the null hypothesis and conditional likelihood).}

\item{parameter.set}{A string indicating the parameterization:
'h2' for proportions of variance (\code{h2_i}) and total variance (s2), or
'sigma2' for direct variance components (s21, s22, ..., s2e).}

\item{TOL}{Numerical tolerance for convergence and constraints (default: 1e-8).}
}
\value{
The likelihood ratio test statistic.
}
\description{
Conducts a specialized likelihood ratio test to test for a specific genetic component
being zero (e.g., h12 = 0). This function is tailored for scenarios where a single
"nuisance" matrix (\code{K2_list[[1]]}) is provided, and it leverages an eigen-decomposition
of this matrix for potentially faster computations. It supports both \code{sigma2} and \code{h2} parameter sets.
}
