\name{defunct Functions}
\alias{varComp.LinScore.LC12}
\alias{varComp.LinScore.LC12Boundary}
\alias{varComp.LinScore.approximation}

\title{	Defunct functions
}
\description{
Functions to be removed from future release. 
}
\usage{
varComp.LinScore.LC12(null.fit, X, Y, K, null, w, ...)
varComp.LinScore.LC12Boundary (null.fit, X, Y, K, null, ...)
varComp.LinScore.approximation(approximation, ...)
}
\arguments{
\item{null.fit,X,Y,K,null,w,approximation,\dots}{To be described later, if indeed needed.}
}
% \keyword{ htest }
