\name{Not Implemented Functions}
\alias{varComp.LinScore.beta}
\alias{varComp.LinScore.cumulant3}
\alias{varComp.LinScore.pboot}
\alias{varComp.LinScore.saddlepoint}

\alias{varComp.RWD88.test}
\alias{varComp.Wald.test}

\title{	Not-yet-implemented functions
}
\description{
Functions to be implemented in future releases. 
}
\usage{
varComp.LinScore.beta(...) 
varComp.LinScore.cumulant3(...) 
varComp.RWD88.test(control,...)
varComp.Wald.test(control,...)
varComp.LinScore.pboot(...)
varComp.LinScore.saddlepoint(...) 
}
\arguments{
\item{control}{list of control parameters.}
\item{\dots}{Place holder.}
}

% \keyword{ htest }
